requirements.txt
Flask==2.2.2
cryptography
mysql-connector-python